requirements.txt
albumentations==2.0.8
matplotlib==3.10.3
numpy==2.3.2
opencv_python==4.10.0.84
opencv_python_headless==4.12.0.88
Pillow==11.3.0
scikit_learn==1.7.1
scipy==1.16.1
segmentation_models_pytorch==0.5.0
torch==2.8.0.dev20250419+cu128
torchvision==0.22.0.dev20250420+cu128
tqdm==4.67.1
